- bugFix: missing check for folder, changing to current kernel version
git-svn-id: https://efaps.googlecode.com/svn/maven-eFaps-PlugIn/trunk@4309 fee104cc-1dfa-8c0f-632d-d3b7e6b59fb0

+++ src/main/java/org/efaps/maven/plugin/install/GenerateInstallationMojo.java
@@ -316,10 +316,12 @@ public class GenerateInstallationMojo
                 final File dstFile = new File(this.targetDirectory, _rootPackage + fileName);
                 FileUtils.copyFile(srcFile, dstFile, true);
             }
-            for (final String fileName : getCopyFiles(getOutputDirectory())) {
-                final File srcFile = new File(getOutputDirectory(), fileName);
-                final File dstFile = new File(this.targetDirectory, _rootPackage + fileName);
-                FileUtils.copyFile(srcFile, dstFile, true);
+            if (getOutputDirectory().exists()) {
+                for (final String fileName : getCopyFiles(getOutputDirectory())) {
+                    final File srcFile = new File(getOutputDirectory(), fileName);
+                    final File dstFile = new File(this.targetDirectory, _rootPackage + fileName);
+                    FileUtils.copyFile(srcFile, dstFile, true);
+                }
             }
         } catch (final IOException e) {
             throw new MojoExecutionException("could not copy files", e);
